// crm-nest-dev/src/components/log-notes/enum/log-note-actions.enum.ts
export enum LogNoteActions {
  CREATE = 1,
  EDIT = 2,
  COMMENT = 3,
  CHANGE_STATUS = 4,
  CHANGE_VALUE = 5,
  CHANGE_DUE_DATE = 6,
  CHANGE_NAME = 7,
  CHANGE_START_DATE = 8,
  CHANGE_CURENCY = 9,
  CHANGE_DELIVERY_DATE = 10,
  CREATE_ORDER_ITEM = 11,
  DELETE_ORDER_ITEM = 13,
  CREATE_INVOICE = 12,
  CREATE_PAYMENT = 14,
  CHANGE_METHOD = 15,
  CHANGE_INVOCE = 16,
  CHANGE_AMOUNT = 17,
  CHANGE_DATE = 18,
  UPLOAD_FILE = 19,
  CREATE_ORDER = 21,
  CREATE_DEAL = 22,
  UPLOAD_FILE_RAW = 23,
  CHANGE_PUBLIC = 24,
  CHANGE_PRIORITY = 25,
  CHANGE_ARCHIVE = 26,
  CREATE_CHECKLIST = 27,
  USER_ASIGN = 28,
  USER_LEFT_TASK = 29,
  DONE_CHECKLIST_ITEM = 30,
  LEFT_CHECKLIST_ITEM = 31,
  REACTION = 32,
  CHANGE_DESCRIPTION = 33,
  CHANGE_PROVINCE = 34,
  CHANGE_INVOICE_CUSTOMER_NAME = 35,
  CHANGE_COUNTRY_NAME = 36,
  CHANGE_GENDER = 37,
  CHANGE_PHONE = 38,
  CHANGE_EMAIL = 39,
  CHANGE_CARD_VISIT = 40,
  CREATE_PAYMENT_PARTNER = 41,
  CHANGE_VAT = 42
}

export enum LogNoteColumnChange {
  CHANGE_STATUS = 'status',
  CHANGE_VALUE = 'value',
  CHANGE_DUE_DATE = 'due date',
  CHANGE_NAME = 'name',
  CHANGE_START_DATE = 'start date',
  CHANGE_CURENCY = 'currency',
  CHANGE_DELIVERY_DATE = 'delivery date',
  CHANGE_METHOD = 'payment method',
  CHANGE_INVOCE = 'invoice',
  CHANGE_AMOUNT = 'amount',
  CHANGE_DATE = 'date',
  CHANGE_PUBLIC = 'security',
  CHANGE_PRIORITY = 'Priority',
  CHANGE_ARCHIVE = 'Archive',
  CREATE_CHECKLIST = 'CheckList',
  USER_ASIGN = 'Join',
  USER_LEFT_TASK = 'Left',
  DONE_CHECKLIST_ITEM = 'Done',
  CHANGE_DESCRIPTION = 'Descriptions',
  CHANGE_PROVINCE = 'Province',
  CHANGE_COUNTRY_NAME = 'Country Name',
  CHANGE_INVOICE_CUSTOMER_NAME = 'Invoice Customer Name',
  CHANGE_GENDER = 'gender',
  CHANGE_PHONE = 'phone',
  CHANGE_EMAIL = 'email',
  CHANGE_CARD_VISIT = 'card visit',
  CHANGE_VAT = 'VAT'
}